refactor: extract shared contracts and storage package
