+++ shared/pyproject.toml
@@ -1,0 +1,15 @@
+[project]
+name = "ohsheet-shared"
+version = "0.1.0"
+description = "Shared contracts and storage protocol for Oh Sheet pipeline"
+requires-python = ">=3.10"
+dependencies = [
+    "pydantic>=2.5",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["shared"]
