+++ shared/shared/__init__.py
@@ -1,0 +1,1 @@
+"""Shared contracts and storage for the Oh Sheet pipeline."""
