+++ shared/shared/storage/__init__.py
@@ -1,0 +1,1 @@
+"""Blob storage protocol and implementations."""
